Can now generate items according to table ID
Currently, I have the game running on a timer, so an attempt to generate an item is done every second. A bonus is applied whenever no item is generated, so that it's easier to get an item on the next roll.

+++ Assets/Scripts/Table.cs
@@ -3,6 +3,6 @@
 [System.Serializable]
 public class Table
 {
-    public float tableId;                //table identification. Each enemy encountered will drop an item from a table with the given ID.
+    public int tableId;                //table identification. Each enemy encountered will drop an item from a table with the given ID.
     public Item[] tableItems;      //list of items contained in the table.
 }
